forntend and backend changes + authorization and roles

+++ app/routers/topic.py
@@ -1,16 +1,20 @@
-from fastapi import APIRouter, HTTPException
+from typing import Annotated
+
+from fastapi import APIRouter, HTTPException, Security
 from fastapi import Depends
 from sqlalchemy.orm import Session
 
 from app.crud import topic as crud_topic
-from app.dependencies import get_db
+from app.dependencies import get_db, auth_scheme
+from app.routers.auth import get_current_user
 from app.schemas.topic import TopicCreate, TopicDB, TopicUpdate
+from app.schemas.user import UserDB
 
-router = APIRouter(dependencies=[Depends(get_db)], tags=["Topic"], prefix="/crud/topic")
+router = APIRouter(dependencies=[Depends(get_db), Security(get_current_user, scopes=["user"])], tags=["Topic"], prefix="/crud/topic")
 
 
 @router.get("/all", response_model=list[TopicDB])
-def get_all_topics(db: Session = Depends(get_db)):
+def get_all_topics(current_user: Annotated[UserDB, Security(get_current_user, scopes=["admin"])], db: Session = Depends(get_db)):
     return crud_topic.get_all_topics_db(db)
 
 
@@ -23,15 +27,20 @@ def get_topic_by_id(topic_id: int, db: Session = Depends(get_db)):
 
 
 @router.post("/add", response_model=TopicDB)
-def create_topic(topic: TopicCreate, db: Session = Depends(get_db)):
+def create_topic(topic: TopicCreate, current_user: Annotated[UserDB, Security(get_current_user, scopes=["user"])], db: Session = Depends(get_db)):
+    topic.user_id = current_user.id
     return crud_topic.create_topic_db(topic, db)
 
 
 @router.delete("/{topic_id}")
-def remove_topic_by_id(topic_id: int, db: Session = Depends(get_db)):
+def remove_topic_by_id(topic_id: int, current_user: Annotated[UserDB, Security(get_current_user, scopes=["user"])], db: Session = Depends(get_db)):
     topic = crud_topic.get_by_topic_id_db(topic_id, db)
     if not topic:
         raise HTTPException(status_code=404, detail="Topic not found, cannot delete")
+    if current_user.is_admin:
+        return crud_topic.delete_topic_by_id_db(topic_id, db)
+    if current_user.id != topic.user_id:
+        raise HTTPException(status_code=501, detail="Unauthorized!!")
     return crud_topic.delete_topic_by_id_db(topic_id, db)
